refactor: update user form
